Fix build

+++ cmd/calibrecli/main.go
@@ -20,9 +20,7 @@ var (
 func must(err error, reason string, callback func()) {
 	if err != nil {
 		fmt.Fprintln(os.Stderr, "Error while", reason)
-		if callback != nil {
-			callback()
-		}
+		if callback != nil { callback() }
 		exitMessage(err)
 	}
 }
@@ -32,14 +30,14 @@ func exitMessage(text interface{}) {
 	os.Exit(1)
 }
 
-func parseArgs() (string, string) {
+func parseArgs() {
 	flag.StringVar(&flagUsername, "username", "", "username for calibre")
 	flag.StringVar(&flagPassword, "password", "", "username for calibre")
 	flag.StringVar(&flagURL, "url", "", "")
 	flag.Parse()
 
-	if v := os.Getenv("CALIBRE_PASSWORD"); v != "" { flagPassword = &v }
-	if v := os.Getenv("CALIBRE_USERNAME"); v != "" { flagUsername = &v }
+	if v := os.Getenv("CALIBRE_PASSWORD"); v != "" { flagPassword = v }
+	if v := os.Getenv("CALIBRE_USERNAME"); v != "" { flagUsername = v }
 	if v := os.Getenv("CALIBRE_URL"); v != "" { flagURL = v }
 
 	if flagURL == "" { must(errors.New("empty URL"), "parsing arguments", nil) }
@@ -72,24 +70,19 @@ func main() {
 			exitMessage("usage: clibrecli list <book|lang|categories|series|authors>")
 		}
 	case "delete":
-		if flag.Arg(1) == "" {
-			exitMessage("usage: clibrecli delete <BOOKID>")
-		}
+		if flag.Arg(1) == "" { exitMessage("usage: clibrecli delete <BOOKID>") }
+
 		id, err := strconv.ParseUint(flag.Arg(1), 10, 0)
 		must(err, "parsing BOOKID", nil)
 
-		if prompt(false, "Delete book") {
-			deleteBook(api, id)
-		}
+		if prompt(false, "Delete book") { deleteBook(api, id) }
 	case "download":
-		if flag.Arg(1) == "" {
-			exitMessage("usage: clibrecli download <URL>")
-		}
-		downloadBook(api, flag.Arg(1))
+		if flag.Arg(1) == "" { exitMessage("usage: clibrecli download <URL>") }
+
+		// TODO
 	case "upload":
-		if flag.Arg(1) == "" {
-			exitMessage("usage: clibrecli upload <FILPATH> [FILEPATH..]")
-		}
+		if flag.Arg(1) == "" { exitMessage("usage: clibrecli upload <FILPATH> [FILEPATH..]") }
+
 		b, err := uploadFile(api, flag.Arg(1), flag.Args()[2:])
 		must(err, "uploading book", func() {
 			if prompt(true, fmt.Sprintf("Delete book: %s (%d)", b.Title, b.ID())) {
